+++ lib/HttpCapture/DevHttpCaptureCompiler/src/main/java/dev/capture/activity/DevHttpCaptureMainActivity.kt
@@ -40,7 +40,7 @@ class DevHttpCaptureMainActivity : BaseDevHttpActivity<DevHttpCaptureMainActivit
             if (!isFinishing) {
                 if (isQuerying) {
                     if (size == 0) {
-                        UtilsCompiler.toastIMPL().normal(
+                        UtilsCompiler.toastImpl().normal(
                             R.string.dev_http_capture_querying
                         )
                     }
@@ -54,7 +54,7 @@ class DevHttpCaptureMainActivity : BaseDevHttpActivity<DevHttpCaptureMainActivit
                     binding.vidRv,
                     binding.vidTipsInclude.vidTipsFl
                 )
-                UtilsCompiler.toastIMPL().success(
+                UtilsCompiler.toastImpl().success(
                     R.string.dev_http_capture_query_complete
                 )
                 // 重置刷新点击处理
@@ -137,7 +137,7 @@ class DevHttpCaptureMainActivity : BaseDevHttpActivity<DevHttpCaptureMainActivit
         ViewHelper.get()
             .setOnClick(object : OnDebouncingClickListener(UtilsCompiler.REFRESH_CLICK) {
                 override fun doClick(view: View) {
-                    UtilsCompiler.toastIMPL().normal(
+                    UtilsCompiler.toastImpl().normal(
                         R.string.dev_http_capture_querying
                     )
                     if (!UtilsCompiler.isQuerying()) {
@@ -148,7 +148,7 @@ class DevHttpCaptureMainActivity : BaseDevHttpActivity<DevHttpCaptureMainActivit
                 }
 
                 override fun doInvalidClick(view: View) {
-                    UtilsCompiler.toastIMPL().normal(
+                    UtilsCompiler.toastImpl().normal(
                         R.string.dev_http_capture_querying
                     )
                 }