+++ lib/DevSimple/CHANGELOG.md
@@ -21,7 +21,7 @@ Version 1.0.1 *(2024-06-04)*
 
 * `[Config]` 更新 IUIController、UIThemeExt 默认配置
 
-* `[Update]` 更新 SimpleActivityIMPL、SimpleFragmentIMPL 匿名函数处理以及相关联实用类变更
+* `[Update]` 更新 SimpleActivityImpl、SimpleFragmentImpl 匿名函数处理以及相关联实用类变更
 
 Version 1.0.0 *(2024-05-15)*
 ----------------------------